Add work on how commands are added and add various commands
To add a new command you call NewCommand with a name, description, set of arguments(used for help text, might use in the function somehow later), and a function to call.
Add roll, help, setPrefix, and messageLifetime commands.
messages from the bot and the message that called it are now deleted after the set messageLifetime. This is to prevent clutter in chat.